+++ internal/pipeline/compiler/transpiler.go
@@ -1014,7 +1014,7 @@ func Transpile(program *ast.Program, a *analyzer.Analyzer, opts TranspileOptions
 	if ctx.usedModules["syscall/js"] {
 		finalBuf.WriteString("import \"syscall/js\"\n")
 	}
-	if ctx.usedModules["page_write"] {
+	if ctx.usedModules["doc_write"] {
 		finalBuf.WriteString("import \"path/filepath\"\n")
 	}